requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
decouple==0.0.7
Deprecated==1.2.18
Django==5.1.2
djangorestframework==3.15.2
fastapi==0.115.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-pubsub==2.29.0
google-cloud-run==0.10.18
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.6.1
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.40.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
wrapt==1.17.2
zipp==3.21.0
